+++ client-config-app/src/lib/env.ts
@@ -7,3 +7,8 @@ export const MAX_NUM = 150;
  * API 地址
  */
 export const BASE_URL = process.env.REACT_APP_API_ROOT || "/api/";
+
+/**
+ * AAD Client ID
+ */
+export const AAD_ID = process.env.REACT_APP_AAD_ID;
