+++ server.net/Program.cs
@@ -38,39 +38,42 @@ builder.Services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
 builder.Services.AddSingleton<IBotFrameworkHttpAdapter, AdapterWithErrorHandler>();
 // Create the bot as a transient. In this case the ASP Controller is expecting an IBot.
 builder.Services.AddTransient<IBot, TeamsMessagingExtensionsBot>();
+
 // Adding Authentication  
 builder.Services.AddAuthorization(options =>
 {
-    options.AddPolicy("EmployeeOnly", policy =>
+    options.AddPolicy("Admin", policy =>
     {
-        policy.RequireClaim("oid");
+        policy.RequireClaim("wids", new string[] { "62e90394-69f5-4237-9190-012177145e10", "69091246-20e8-4a56-aa4d-066075b2a7a8" });
         policy.AddAuthenticationSchemes(JwtBearerDefaults.AuthenticationScheme);
-        policy.AddRequirements(new AuthorizationRequirement("EmployeeOnly", JwtBearerDefaults.AuthenticationScheme));
+        policy.AddRequirements(new AuthorizationRequirement("AdminOnly", JwtBearerDefaults.AuthenticationScheme));
     });
 });
 builder.Services
     .AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
     .AddMicrosoftIdentityWebApi(builder.Configuration, ConfigKeys.AAD_SECTION);
-// .AddJwtBearer(o =>builder.Configuration
-// {
-//     o.TokenValidationParameters = new TokenValidationParameters()
-//     {
-
-//         // IssuerValidator = (string issuer, SecurityToken securityToken, TokenValidationParameters validationParameters) => "true",
-//         ValidAudiences = builder.Configuration[ConfigKeys.AAD_APP_ID].Split(","),
-//         ValidateAudience = true,
-//         ValidateIssuerSigningKey = true,
-//         ValidateIssuer = true,
-//     };
-// });
 builder.Services.AddSingleton<IAuthorizationHandler, AuthorizationHandler>();
+
 builder.Services.AddMemoryCache();
 
 // Add http services to the container.
 builder.Services.AddHttpClient();
 
+if (builder.Environment.IsDevelopment())
+{
+    builder.Services.AddCors();
+}
+
+
 var app = builder.Build();
 
+if (builder.Environment.IsDevelopment())
+{
+    app.UseCors(x => x
+          .AllowAnyOrigin()
+          .AllowAnyMethod()
+          .AllowAnyHeader()); // configure CORS
+}
 
 // Configure the HTTP request pipeline.
